Fix scene creation dialogue
Somehow the lights selected on the scene creation dialogue were not
beint recorded properly. I'm sure they worked at some point, but I don't
see how it's possible based on the code. An attempt was being made to
pull the checkbox flag from the model, rather than the delegates. This
has now been changed so that the delegates are correctly tested instead.