Allow timings to be applied to light/group changes
Adds several new methods for controlling the brightness, colour, etc. of
lights and groups which allow timings to be provided for the duration of
the transition.

This is used in the Sailfish app when changing light details dynamically
using the colour picker, since changes are now made immediately to
prevent feedback cycles from causing problems with the selections.